requirements.txt
ujson
psycopg2-binary
click
